Fix searchWindow type to Long to match searchWindowUsed

+++ src/ext/java/org/opentripplanner/ext/legacygraphqlapi/datafetchers/LegacyGraphQLQueryTypeImpl.java
@@ -589,7 +589,7 @@ public class LegacyGraphQLQueryTypeImpl
 
       callWith.argument("wheelchair", request::setWheelchairAccessible);
       callWith.argument("numItineraries", request::setNumItineraries);
-      callWith.argument("searchWindow", (Integer m) -> request.searchWindow = Duration.ofSeconds(m));
+      callWith.argument("searchWindow", (Long m) -> request.searchWindow = Duration.ofSeconds(m));
       callWith.argument("maxWalkDistance", request::setMaxWalkDistance);
       // callWith.argument("maxSlope", request::setMaxSlope);
       callWith.argument("maxPreTransitTime", request::setMaxPreTransitTime);
